Move tests to tests.rs
