+++ src/routes/trays/mod.rs
@@ -1,6 +1,5 @@
 pub mod models;
 pub mod services;
-pub mod views;
-
 #[cfg(test)]
-pub mod tests;
+mod tests;
+pub mod views;
