+++ src/routes/trays/tests.rs
@@ -46,10 +46,10 @@ async fn test_tray_crud_operations() {
         .unwrap();
 
     let (status, body) = extract_response_body(response).await;
-    
+
     if status == StatusCode::CREATED {
         println!("✅ Tray creation successful");
-        
+
         // Validate response structure
         assert!(body["id"].is_string(), "Response should include ID");
         assert!(body["name"].as_str().unwrap().contains("Test Tray CRUD"));
@@ -78,7 +78,12 @@ async fn test_tray_crud_operations() {
         if get_status == StatusCode::OK {
             println!("✅ Tray retrieval successful");
             assert_eq!(get_body["id"], tray_id);
-            assert!(get_body["name"].as_str().unwrap().contains("Test Tray CRUD"));
+            assert!(
+                get_body["name"]
+                    .as_str()
+                    .unwrap()
+                    .contains("Test Tray CRUD")
+            );
         } else {
             println!("⚠️  Tray retrieval failed: {get_status}");
         }
@@ -134,12 +139,13 @@ async fn test_tray_crud_operations() {
         } else {
             println!("📋 Tray delete returned: {delete_status}");
         }
-        
     } else {
         println!("⚠️  Tray creation failed: Status {status}, Body: {body}");
         // Document the current behavior even if it fails
-        assert!(status.is_client_error() || status.is_server_error(),
-               "Tray creation should either succeed or fail gracefully");
+        assert!(
+            status.is_client_error() || status.is_server_error(),
+            "Tray creation should either succeed or fail gracefully"
+        );
     }
 }
 
@@ -161,28 +167,39 @@ async fn test_tray_list_operations() {
         .unwrap();
 
     let (list_status, list_body) = extract_response_body(list_response).await;
-    
+
     if list_status == StatusCode::OK {
         println!("✅ Tray listing successful");
         assert!(list_body.is_array(), "Trays list should be an array");
         let trays = list_body.as_array().unwrap();
         println!("Found {} trays in the system", trays.len());
-        
+
         // Validate structure of trays in list
         for tray in trays {
             assert!(tray["id"].is_string(), "Each tray should have ID");
-            assert!(tray["created_at"].is_string(), "Each tray should have created_at");
-            assert!(tray["last_updated"].is_string(), "Each tray should have last_updated");
-            
+            assert!(
+                tray["created_at"].is_string(),
+                "Each tray should have created_at"
+            );
+            assert!(
+                tray["last_updated"].is_string(),
+                "Each tray should have last_updated"
+            );
+
             // Name can be null, so check if it's present and valid when not null
             if !tray["name"].is_null() {
-                assert!(tray["name"].is_string(), "Tray name should be string when present");
+                assert!(
+                    tray["name"].is_string(),
+                    "Tray name should be string when present"
+                );
             }
         }
     } else {
         println!("⚠️  Tray listing failed: Status {list_status}");
-        assert!(list_status.is_client_error() || list_status.is_server_error(),
-               "Tray listing should either succeed or fail gracefully");
+        assert!(
+            list_status.is_client_error() || list_status.is_server_error(),
+            "Tray listing should either succeed or fail gracefully"
+        );
     }
 }
 
@@ -212,7 +229,7 @@ async fn test_tray_validation() {
         .unwrap();
 
     let (status, _body) = extract_response_body(response).await;
-    
+
     if status.is_client_error() {
         println!("✅ Tray validation working - rejected negative axis value");
     } else if status == StatusCode::CREATED {
@@ -243,7 +260,7 @@ async fn test_tray_validation() {
         .unwrap();
 
     let (status, _body) = extract_response_body(response).await;
-    
+
     if status.is_client_error() {
         println!("✅ Tray validation rejects zero dimensions");
     } else if status == StatusCode::CREATED {
@@ -261,11 +278,11 @@ async fn test_tray_filtering_and_sorting() {
     let test_trays = [
         ("96-Well Plate", 12, 8),
         ("384-Well Plate", 24, 16),
-        ("48-Well Plate", 8, 6)
+        ("48-Well Plate", 8, 6),
     ];
 
     let mut created_ids = Vec::new();
-    
+
     for (name, x_axis, y_axis) in test_trays {
         let tray_data = json!({
             "name": format!("{} {}", name, &uuid::Uuid::new_v4().to_string()[..8]),
@@ -310,12 +327,15 @@ async fn test_tray_filtering_and_sorting() {
             .unwrap();
 
         let (filter_status, filter_body) = extract_response_body(filter_response).await;
-        
+
         if filter_status == StatusCode::OK {
             println!("✅ Tray filtering endpoint accessible");
             let filtered_trays = filter_body.as_array().unwrap();
-            println!("Filtered trays by name=96-Well Plate: {} results", filtered_trays.len());
-            
+            println!(
+                "Filtered trays by name=96-Well Plate: {} results",
+                filtered_trays.len()
+            );
+
             // Check if filtering actually works
             let mut filtering_works = true;
             for tray in filtered_trays {
@@ -326,7 +346,7 @@ async fn test_tray_filtering_and_sorting() {
                     }
                 }
             }
-            
+
             if filtering_works && !filtered_trays.is_empty() {
                 println!("✅ Tray filtering appears to work correctly");
             } else if filtered_trays.is_empty() {
@@ -350,7 +370,7 @@ async fn test_tray_filtering_and_sorting() {
             .unwrap();
 
         let (sort_status, _) = extract_response_body(sort_response).await;
-        
+
         if sort_status == StatusCode::OK {
             println!("✅ Tray sorting endpoint accessible");
         } else {
@@ -378,7 +398,11 @@ async fn test_tray_not_found() {
         .unwrap();
 
     let (status, _body) = extract_response_body(response).await;
-    assert_eq!(status, StatusCode::NOT_FOUND, "Should return 404 for non-existent tray");
+    assert_eq!(
+        status,
+        StatusCode::NOT_FOUND,
+        "Should return 404 for non-existent tray"
+    );
     println!("✅ Tray 404 handling working correctly");
 }
 
@@ -432,13 +456,18 @@ async fn test_tray_configuration_crud_operations() {
         .unwrap();
 
     let (status, body) = extract_response_body(response).await;
-    
+
     if status == StatusCode::CREATED {
         println!("✅ Tray configuration creation successful");
-        
+
         // Validate response structure
         assert!(body["id"].is_string(), "Response should include ID");
-        assert!(body["name"].as_str().unwrap().contains("Test Tray Config CRUD"));
+        assert!(
+            body["name"]
+                .as_str()
+                .unwrap()
+                .contains("Test Tray Config CRUD")
+        );
         assert_eq!(body["experiment_default"], false);
         assert!(body["created_at"].is_string());
         assert!(body["last_updated"].is_string());
@@ -447,10 +476,10 @@ async fn test_tray_configuration_crud_operations() {
         if body["trays"].is_array() {
             let trays = body["trays"].as_array().unwrap();
             println!("   Configuration has {} tray assignments", trays.len());
-            
+
             if trays.len() == 2 {
                 println!("   ✅ All tray assignments preserved");
-                
+
                 // Validate assignment structure
                 for assignment in trays {
                     assert!(assignment["order_sequence"].is_number());
@@ -479,7 +508,7 @@ async fn test_tray_configuration_crud_operations() {
         if get_status == StatusCode::OK {
             println!("✅ Tray configuration retrieval successful");
             assert_eq!(get_body["id"], tray_config_id);
-            
+
             // Validate related data structure
             if get_body["trays"].is_array() {
                 println!("   ✅ Trays array present and loaded");
@@ -540,12 +569,13 @@ async fn test_tray_configuration_crud_operations() {
         } else {
             println!("📋 Tray configuration delete returned: {delete_status}");
         }
-        
     } else {
         println!("⚠️  Tray configuration creation failed: Status {status}, Body: {body}");
         // Document the current behavior even if it fails
-        assert!(status.is_client_error() || status.is_server_error(),
-               "Tray configuration creation should either succeed or fail gracefully");
+        assert!(
+            status.is_client_error() || status.is_server_error(),
+            "Tray configuration creation should either succeed or fail gracefully"
+        );
     }
 }
 
@@ -567,24 +597,41 @@ async fn test_tray_configuration_list_operations() {
         .unwrap();
 
     let (list_status, list_body) = extract_response_body(list_response).await;
-    
+
     if list_status == StatusCode::OK {
         println!("✅ Tray configuration listing successful");
-        assert!(list_body.is_array(), "Tray configurations list should be an array");
+        assert!(
+            list_body.is_array(),
+            "Tray configurations list should be an array"
+        );
         let tray_configs = list_body.as_array().unwrap();
-        println!("Found {} tray configurations in the system", tray_configs.len());
-        
+        println!(
+            "Found {} tray configurations in the system",
+            tray_configs.len()
+        );
+
         // Validate structure of tray configurations in list
         for config in tray_configs {
             assert!(config["id"].is_string(), "Each config should have ID");
-            assert!(config["created_at"].is_string(), "Each config should have created_at");
-            assert!(config["last_updated"].is_string(), "Each config should have last_updated");
-            assert!(config["experiment_default"].is_boolean(), "Each config should have experiment_default");
+            assert!(
+                config["created_at"].is_string(),
+                "Each config should have created_at"
+            );
+            assert!(
+                config["last_updated"].is_string(),
+                "Each config should have last_updated"
+            );
+            assert!(
+                config["experiment_default"].is_boolean(),
+                "Each config should have experiment_default"
+            );
         }
     } else {
         println!("⚠️  Tray configuration listing failed: Status {list_status}");
-        assert!(list_status.is_client_error() || list_status.is_server_error(),
-               "Tray configuration listing should either succeed or fail gracefully");
+        assert!(
+            list_status.is_client_error() || list_status.is_server_error(),
+            "Tray configuration listing should either succeed or fail gracefully"
+        );
     }
 }
 
@@ -626,13 +673,13 @@ async fn test_tray_configuration_default_behavior() {
         .unwrap();
 
     let (status, body) = extract_response_body(response).await;
-    
+
     if status == StatusCode::CREATED {
         println!("✅ Default tray configuration created successfully");
         assert_eq!(body["experiment_default"], true);
-        
+
         let first_config_id = body["id"].as_str().unwrap();
-        
+
         // Create another default configuration - should set the first one to false
         let second_default_config_data = json!({
             "name": format!("Second Default Config {}", uuid::Uuid::new_v4()),
@@ -667,11 +714,11 @@ async fn test_tray_configuration_default_behavior() {
             .unwrap();
 
         let (second_status, second_body) = extract_response_body(second_response).await;
-        
+
         if second_status == StatusCode::CREATED {
             println!("✅ Second default configuration created");
             assert_eq!(second_body["experiment_default"], true);
-            
+
             // Check if the first configuration is no longer default
             let check_first_response = app
                 .clone()
@@ -686,12 +733,16 @@ async fn test_tray_configuration_default_behavior() {
                 .unwrap();
 
             let (check_status, check_body) = extract_response_body(check_first_response).await;
-            
+
             if check_status == StatusCode::OK {
                 if check_body["experiment_default"] == false {
-                    println!("   ✅ Default configuration behavior working - first config no longer default");
+                    println!(
+                        "   ✅ Default configuration behavior working - first config no longer default"
+                    );
                 } else {
-                    println!("   📋 Multiple default configurations allowed (or default logic not working)");
+                    println!(
+                        "   📋 Multiple default configurations allowed (or default logic not working)"
+                    );
                 }
             }
         } else {
@@ -737,7 +788,7 @@ async fn test_tray_dimensions_validation() {
             .unwrap();
 
         let (status, _body) = extract_response_body(response).await;
-        
+
         if status == StatusCode::CREATED {
             println!("✅ Tray accepts {description} ({x_axis}x{y_axis})");
         } else {
@@ -814,20 +865,20 @@ async fn test_tray_configuration_complex_structure() {
         .unwrap();
 
     let (status, body) = extract_response_body(response).await;
-    
+
     if status == StatusCode::CREATED {
         println!("✅ Complex tray configuration created successfully");
-        
+
         let config_id = body["id"].as_str().unwrap();
-        
+
         // Validate complex structure
         if body["trays"].is_array() {
             let assignments = body["trays"].as_array().unwrap();
             println!("   Configuration has {} assignments", assignments.len());
-            
+
             if assignments.len() == 3 {
                 println!("   ✅ All assignments preserved");
-                
+
                 // Count total trays across all assignments
                 let mut total_trays = 0;
                 for assignment in assignments {
@@ -835,18 +886,19 @@ async fn test_tray_configuration_complex_structure() {
                         total_trays += assignment["trays"].as_array().unwrap().len();
                     }
                 }
-                
-                if total_trays == 4 {  // 2 + 1 + 1 trays across assignments
+
+                if total_trays == 4 {
+                    // 2 + 1 + 1 trays across assignments
                     println!("   ✅ All {total_trays} trays preserved across assignments");
                 }
-                
+
                 // Check order sequence sorting
                 let mut sequences: Vec<i64> = Vec::new();
                 for assignment in assignments {
                     sequences.push(assignment["order_sequence"].as_i64().unwrap());
                 }
                 let is_sorted = sequences.windows(2).all(|w| w[0] <= w[1]);
-                
+
                 if is_sorted {
                     println!("   ✅ Assignments are properly sorted by order_sequence");
                 } else {
@@ -854,7 +906,7 @@ async fn test_tray_configuration_complex_structure() {
                 }
             }
         }
-        
+
         // Test retrieval of complex configuration
         let get_response = app
             .clone()
@@ -869,10 +921,10 @@ async fn test_tray_configuration_complex_structure() {
             .unwrap();
 
         let (get_status, get_body) = extract_response_body(get_response).await;
-        
+
         if get_status == StatusCode::OK {
             println!("   ✅ Complex configuration retrieval successful");
-            
+
             // Validate that all data is loaded correctly
             if get_body["trays"].is_array() {
                 let assignments = get_body["trays"].as_array().unwrap();
@@ -881,7 +933,6 @@ async fn test_tray_configuration_complex_structure() {
                 }
             }
         }
-        
     } else {
         println!("📋 Complex tray configuration creation failed: Status {status}");
     }
@@ -906,7 +957,11 @@ async fn test_tray_configuration_not_found() {
         .unwrap();
 
     let (status, _body) = extract_response_body(response).await;
-    assert_eq!(status, StatusCode::NOT_FOUND, "Should return 404 for non-existent tray configuration");
+    assert_eq!(
+        status,
+        StatusCode::NOT_FOUND,
+        "Should return 404 for non-existent tray configuration"
+    );
     println!("✅ Tray configuration 404 handling working correctly");
 }
 
@@ -916,7 +971,7 @@ async fn test_tray_workflow_comprehensive() {
 
     println!("📋 TRAY COMPREHENSIVE WORKFLOW TEST");
     println!("   Testing complete tray and tray configuration lifecycle");
-    
+
     // Step 1: Create individual trays
     let individual_tray_data = json!({
         "name": format!("Workflow Individual Tray {}", uuid::Uuid::new_v4()),
@@ -939,13 +994,13 @@ async fn test_tray_workflow_comprehensive() {
         .unwrap();
 
     let (individual_status, _individual_body) = extract_response_body(individual_response).await;
-    
+
     if individual_status == StatusCode::CREATED {
         println!("   ✅ Step 1: Individual tray created successfully");
     } else {
         println!("   📋 Step 1: Individual tray creation returned: {individual_status}");
     }
-    
+
     // Step 2: Create comprehensive tray configuration
     let comprehensive_config_data = json!({
         "name": format!("Comprehensive Workflow Config {}", uuid::Uuid::new_v4()),
@@ -992,12 +1047,12 @@ async fn test_tray_workflow_comprehensive() {
         .unwrap();
 
     let (config_status, config_body) = extract_response_body(config_response).await;
-    
+
     if config_status == StatusCode::CREATED {
         println!("   ✅ Step 2: Comprehensive tray configuration created");
-        
+
         let config_id = config_body["id"].as_str().unwrap();
-        
+
         // Step 3: Validate configuration retrieval and structure
         let get_config_response = app
             .clone()
@@ -1012,33 +1067,461 @@ async fn test_tray_workflow_comprehensive() {
             .unwrap();
 
         let (get_config_status, get_config_body) = extract_response_body(get_config_response).await;
-        
+
         if get_config_status == StatusCode::OK {
             println!("   ✅ Step 3: Configuration retrieval successful");
-            
+
             // Validate comprehensive structure
             assert_eq!(get_config_body["experiment_default"], true);
-            
+
             if get_config_body["trays"].is_array() {
                 let assignments = get_config_body["trays"].as_array().unwrap();
-                println!("   Configuration loaded with {} assignments", assignments.len());
-                
+                println!(
+                    "   Configuration loaded with {} assignments",
+                    assignments.len()
+                );
+
                 if assignments.len() == 2 {
                     println!("   ✅ Step 4: All assignments preserved and loaded");
                 }
             }
-            
+
             if get_config_body["associated_experiments"].is_array() {
                 println!("   ✅ Step 5: Associated experiments structure present");
             }
         }
-        
+
         println!("   📋 Comprehensive workflow test completed successfully");
-        
     } else {
         println!("   ⚠️  Comprehensive workflow test failed - config creation: {config_status}");
     }
-    
+
     // This test always passes - it's for comprehensive workflow documentation
     // Documents comprehensive tray workflow behavior
+}
+
+async fn create_test_tray_via_api(app: &axum::Router) -> Result<Value, Box<dyn std::error::Error>> {
+    let tray_data = json!({
+        "name": "TestTray",
+        "experiment_default": true,
+        "trays": [
+            {
+                "trays": [
+                    {
+                        "name": "P1",
+                        "qty_x_axis": 8,
+                        "qty_y_axis": 12,
+                        "well_relative_diameter": 0.6
+                    }
+                ],
+                "rotation_degrees": 0,
+                "order_sequence": 1
+            },
+            {
+                "trays": [
+                    {
+                        "name": "P2",
+                        "qty_x_axis": 8,
+                        "qty_y_axis": 12,
+                        "well_relative_diameter": 0.6
+                    }
+                ],
+                "rotation_degrees": 180,
+                "order_sequence": 2
+            }
+        ]
+    });
+
+    let response = app
+        .clone()
+        .oneshot(
+            Request::builder()
+                .method("POST")
+                .uri("/api/trays")
+                .header("content-type", "application/json")
+                .body(Body::from(tray_data.to_string()))
+                .unwrap(),
+        )
+        .await
+        .unwrap();
+
+    let (status, body) = extract_response_body(response).await;
+    if status == StatusCode::CREATED {
+        Ok(body)
+    } else {
+        Err(format!("Failed to create tray: {status:?}, body: {body:?}").into())
+    }
+}
+
+#[tokio::test]
+async fn test_create_tray() {
+    let app = setup_test_app().await;
+
+    let tray_data = json!({
+        "name": "TestTray",
+        "experiment_default": true,
+        "trays": [
+            {
+                "trays": [
+                    {
+                        "name": "P1",
+                        "qty_x_axis": 8,
+                        "qty_y_axis": 12,
+                        "well_relative_diameter": 0.6
+                    }
+                ],
+                "rotation_degrees": 0,
+                "order_sequence": 1
+            },
+            {
+                "trays": [
+                    {
+                        "name": "P2",
+                        "qty_x_axis": 8,
+                        "qty_y_axis": 12,
+                        "well_relative_diameter": 0.6
+                    }
+                ],
+                "rotation_degrees": 180,
+                "order_sequence": 2
+            }
+        ]
+    });
+
+    let response = app
+        .clone()
+        .oneshot(
+            Request::builder()
+                .method("POST")
+                .uri("/api/trays")
+                .header("content-type", "application/json")
+                .body(Body::from(tray_data.to_string()))
+                .unwrap(),
+        )
+        .await
+        .unwrap();
+
+    let (status, body) = extract_response_body(response).await;
+    assert_eq!(
+        status,
+        StatusCode::CREATED,
+        "Failed to create tray: {body:?}"
+    );
+
+    // Validate response structure matches expected format
+    assert!(body["id"].is_string());
+    assert_eq!(body["name"], "TestTray");
+    assert_eq!(body["experiment_default"], true);
+    assert!(body["created_at"].is_string());
+    assert!(body["last_updated"].is_string());
+    assert!(body["trays"].is_array());
+    assert_eq!(body["trays"].as_array().unwrap().len(), 2);
+    assert!(body["associated_experiments"].is_array());
+
+    // Validate tray structure
+    let first_tray = &body["trays"][0];
+    assert_eq!(first_tray["order_sequence"], 1);
+    assert_eq!(first_tray["rotation_degrees"], 0);
+    assert_eq!(first_tray["trays"][0]["name"], "P1");
+
+    let second_tray = &body["trays"][1];
+    assert_eq!(second_tray["order_sequence"], 2);
+    assert_eq!(second_tray["rotation_degrees"], 180);
+    assert_eq!(second_tray["trays"][0]["name"], "P2");
+}
+
+#[tokio::test]
+async fn test_get_tray_by_id() {
+    let app = setup_test_app().await;
+
+    // Create a tray first
+    let tray = create_test_tray_via_api(&app)
+        .await
+        .expect("Failed to create test tray");
+    let tray_id = tray["id"].as_str().unwrap();
+
+    // Get the tray by ID
+    let response = app
+        .clone()
+        .oneshot(
+            Request::builder()
+                .method("GET")
+                .uri(format!("/api/trays/{tray_id}"))
+                .body(Body::empty())
+                .unwrap(),
+        )
+        .await
+        .unwrap();
+
+    let (status, body) = extract_response_body(response).await;
+    assert_eq!(status, StatusCode::OK, "Failed to get tray: {body:?}");
+    assert_eq!(body["id"], tray_id);
+    assert_eq!(body["name"], "TestTray");
+    assert_eq!(body["experiment_default"], true);
+}
+
+#[tokio::test]
+async fn test_list_trays() {
+    let app = setup_test_app().await;
+
+    // Create a few trays
+    for i in 1..=3 {
+        let tray_data = json!({
+            "name": format!("TestTray{}", i),
+            "experiment_default": false,
+            "trays": []
+        });
+
+        let response = app
+            .clone()
+            .oneshot(
+                Request::builder()
+                    .method("POST")
+                    .uri("/api/trays")
+                    .header("content-type", "application/json")
+                    .body(Body::from(tray_data.to_string()))
+                    .unwrap(),
+            )
+            .await
+            .unwrap();
+
+        let (status, _) = extract_response_body(response).await;
+        assert_eq!(status, StatusCode::CREATED);
+    }
+
+    // List all trays
+    let response = app
+        .clone()
+        .oneshot(
+            Request::builder()
+                .method("GET")
+                .uri("/api/trays")
+                .body(Body::empty())
+                .unwrap(),
+        )
+        .await
+        .unwrap();
+
+    let (status, body) = extract_response_body(response).await;
+    assert_eq!(status, StatusCode::OK, "Failed to list trays: {body:?}");
+    assert!(body.is_array(), "Response should be an array");
+    let items = body.as_array().unwrap();
+    assert!(items.len() >= 3, "Should have at least 3 trays");
+}
+
+#[tokio::test]
+async fn test_update_tray() {
+    let app = setup_test_app().await;
+
+    // Create a tray first
+    let tray = create_test_tray_via_api(&app)
+        .await
+        .expect("Failed to create test tray");
+    let tray_id = tray["id"].as_str().unwrap();
+
+    // Update the tray - use same format as create
+    let update_data = json!({
+        "name": "UpdatedTestTray",
+        "experiment_default": false,
+        "trays": [
+            {
+                "trays": [
+                    {
+                        "name": "UpdatedP1",
+                        "qty_x_axis": 10,
+                        "qty_y_axis": 14,
+                        "well_relative_diameter": 0.8
+                    }
+                ],
+                "rotation_degrees": 90,
+                "order_sequence": 1
+            }
+        ]
+    });
+
+    let response = app
+        .clone()
+        .oneshot(
+            Request::builder()
+                .method("PUT")
+                .uri(format!("/api/trays/{tray_id}"))
+                .header("content-type", "application/json")
+                .body(Body::from(update_data.to_string()))
+                .unwrap(),
+        )
+        .await
+        .unwrap();
+
+    let (status, body) = extract_response_body(response).await;
+
+    // If PUT doesn't work, try PATCH
+    if status == StatusCode::OK {
+        assert_eq!(body["name"], "UpdatedTestTray");
+        assert_eq!(body["experiment_default"], false);
+        assert_eq!(body["trays"].as_array().unwrap().len(), 1);
+    } else {
+        let response = app
+            .clone()
+            .oneshot(
+                Request::builder()
+                    .method("PATCH")
+                    .uri(format!("/api/trays/{tray_id}"))
+                    .header("content-type", "application/json")
+                    .body(Body::from(update_data.to_string()))
+                    .unwrap(),
+            )
+            .await
+            .unwrap();
+
+        let (status, body) = extract_response_body(response).await;
+        assert_eq!(status, StatusCode::OK, "Failed to update tray: {body:?}");
+        assert_eq!(body["name"], "UpdatedTestTray");
+        assert_eq!(body["experiment_default"], false);
+        assert_eq!(body["trays"].as_array().unwrap().len(), 1);
+    }
+}
+
+#[tokio::test]
+async fn test_delete_tray() {
+    let app = setup_test_app().await;
+
+    // Create a tray first
+    let tray = create_test_tray_via_api(&app)
+        .await
+        .expect("Failed to create test tray");
+    let tray_id = tray["id"].as_str().unwrap();
+
+    // Delete the tray
+    let response = app
+        .clone()
+        .oneshot(
+            Request::builder()
+                .method("DELETE")
+                .uri(format!("/api/trays/{tray_id}"))
+                .body(Body::empty())
+                .unwrap(),
+        )
+        .await
+        .unwrap();
+
+    let (status, _) = extract_response_body(response).await;
+    assert_eq!(status, StatusCode::NO_CONTENT, "Failed to delete tray");
+
+    // Verify it's actually deleted
+    let get_response = app
+        .clone()
+        .oneshot(
+            Request::builder()
+                .method("GET")
+                .uri(format!("/api/trays/{tray_id}"))
+                .body(Body::empty())
+                .unwrap(),
+        )
+        .await
+        .unwrap();
+
+    let (get_status, _) = extract_response_body(get_response).await;
+    assert_eq!(get_status, StatusCode::NOT_FOUND, "Tray should be deleted");
+}
+
+#[tokio::test]
+async fn test_experiment_default_exclusivity() {
+    let app = setup_test_app().await;
+
+    // Create first tray as experiment_default
+    let first_tray_data = json!({
+        "name": "FirstDefault",
+        "experiment_default": true,
+        "trays": []
+    });
+
+    let response = app
+        .clone()
+        .oneshot(
+            Request::builder()
+                .method("POST")
+                .uri("/api/trays")
+                .header("content-type", "application/json")
+                .body(Body::from(first_tray_data.to_string()))
+                .unwrap(),
+        )
+        .await
+        .unwrap();
+
+    let (status, body) = extract_response_body(response).await;
+    assert_eq!(status, StatusCode::CREATED);
+    let first_tray_id = body["id"].as_str().unwrap();
+
+    // Create second tray as experiment_default
+    let second_tray_data = json!({
+        "name": "SecondDefault",
+        "experiment_default": true,
+        "trays": []
+    });
+
+    let response = app
+        .clone()
+        .oneshot(
+            Request::builder()
+                .method("POST")
+                .uri("/api/trays")
+                .header("content-type", "application/json")
+                .body(Body::from(second_tray_data.to_string()))
+                .unwrap(),
+        )
+        .await
+        .unwrap();
+
+    let (status, _) = extract_response_body(response).await;
+    assert_eq!(status, StatusCode::CREATED);
+
+    // Verify first tray is no longer experiment_default
+    let get_response = app
+        .clone()
+        .oneshot(
+            Request::builder()
+                .method("GET")
+                .uri(format!("/api/trays/{first_tray_id}"))
+                .body(Body::empty())
+                .unwrap(),
+        )
+        .await
+        .unwrap();
+
+    let (get_status, get_body) = extract_response_body(get_response).await;
+    assert_eq!(get_status, StatusCode::OK);
+    assert_eq!(
+        get_body["experiment_default"], false,
+        "First tray should no longer be experiment_default"
+    );
+}
+
+#[tokio::test]
+async fn test_validation_errors() {
+    let app = setup_test_app().await;
+
+    // Test with missing required fields
+    let invalid_data = json!({
+        "name": "Invalid Tray"
+        // Missing experiment_default
+    });
+
+    let response = app
+        .clone()
+        .oneshot(
+            Request::builder()
+                .method("POST")
+                .uri("/api/trays")
+                .header("content-type", "application/json")
+                .body(Body::from(invalid_data.to_string()))
+                .unwrap(),
+        )
+        .await
+        .unwrap();
+
+    let (status, _) = extract_response_body(response).await;
+    assert!(
+        status.is_client_error(),
+        "Should reject tray with missing required fields"
+    );
 }
